Efficient checks for stable modules
With a large number of modules in a home package
(in my case 5000) the costs of linear lookups becomes significant.
This changes them to efficient IntMap lookups.

It reduces the cost of `:reload` on unchanged source
from 5.77s to 1.62s on my test case.

I could go further and make `Linker.unload` also take a set,
but I prefer to concentrate on one thing at a time.

Test Plan: harbormaster

Reviewers: simonmar, austin, bgamari

Reviewed By: bgamari

Subscribers: rwbarton, thomie

Differential Revision: https://phabricator.haskell.org/D3611