Add superclasses to quantified constraints
This patch adds suppport for superclasses to quantified constraints.
For example (contrived):

  f :: (forall a. Ord a => Ord (m a)) => m a -> m a -> Bool
  f x y = x==y

Here we need (Eq (m a)); but the quantifed constraint deals only
with Ord.  But we can make it work by using its superclass.

This behaviour finally delivers on the promise of comment:30 of
Trac #9123: we can write an implication constraint that solves
the problem of higher-kinded roles.  Test
quantified-constraints/T8123 demonstrates this in action.